requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
